Convert streamer providers to be called over gRPC
Summary:
Originally, the plugin registered a set of streamer provider impls. Now, the plugin registers the same set of *remote* streamer providers, which forward to servicer providers attached to named services.

This diff converts the base and lte plugins (the only plugins that expose streamers) -- adding the required LTE service (exposing the LTE stream providers) is next in the stack.

The relation of streamer streams to services is n-1 -- that is, a single service can expose as many streamer streams as desired.

See streamer/doc.go (below) for more info.

Reviewed By: xjtian

Differential Revision: D22305991

fbshipit-source-id: a898dea67550687542e9eebe956e980576d88198