Fixed incorrect PLMN and TAC test cases (#1917)
Summary:
## Title
Fixed incorrect PLMN and TAC test cases

There were 2 test incorrect PLMN & TAC cases, which were not executing, although they were part of sanity suite. The reason was that it was not wrapped under any test function. Even after formatting the test cases to be executed in current test framework, the PLMN value was set as 000000 during S1-Setup request. The reason was that the value sent as array of digits was not getting parsed in string format by the S1APTester. This PR fixes these issues.

Pull Request resolved: https://github.com/facebookincubator/magma/pull/1917

Test Plan: Tested with Sanity and verified for zero memory leaks.

Reviewed By: ssanadhya

Differential Revision: D22444025

Pulled By: ulaskozat

fbshipit-source-id: efef213ba7edb14aa5e203299e23679501688c1d