Replace ITTI messaging with ZMQ
Summary:
ZMQ replacing ITTI messaging stack

This is the main diff in the stack

This diff uses ZMQ sockets for ipc communication instead of using ITTI
We use a non-broker based solution. Each task has a PULL socket for receiving all IPC and a number of PUSH sockets depending on how many tasks it needs to communicate with.

We use zloop to manage the sockets as it also provides timer support that we will use later to get rid of ITTI all together.

Main thread communicates with all tasks for termination. But it does not need a PULL socket.

ITTI library is still used to manage timers and signals

Reviewed By: ulaskozat

Differential Revision: D22347710

fbshipit-source-id: 0082fda651adc8995ca82b35b9676f64c5cb25b6