Update magmad.yml docker upgrader to not use proxy
Summary:
Currently, proxying upgrades through the orc8r doesn't work as
the forward proxy has been removed when we migrated from nghttpx to nginx.

We are still figuring out if we need to add this back or can just
remove this configuration all together, but for now set the default
so that upgrades work properly.

Reviewed By: xjtian

Differential Revision: D22519172

fbshipit-source-id: a4f0e9de0e5ffc184817d07ade6eafcb0b56b11c