Add grpc connection reuse to CWF operator (#1950)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1950

This diff is a result of issues seen duing CWAG HA
SIT testing. Previously, we created a new connection for
each client call to the gateway health service. This has
shown to be resource intensive which is problematic for
a variety of reasons (i.e. wasted k8s control plane time,
longer failovers, etc.)

This diff adds a service connection registry by using
the existing registry in orc8r/lib. Since k8s services
are dynamic and external, here we add services on the
fly.

We set service name to svc:port rather than just
svc as in theory a gw port could change during runtime.
This ensures that if this ever occurs, we won't use
the old connection that no longer works.

Reviewed By: xjtian

Differential Revision: D22451948

fbshipit-source-id: 965c87bdc87f2c859f74c5bd7c2680533e9b4f71