Attach lte obsidian handlers to individual lte services
Summary:
This diff removes the obsidian handlers from the lte plugin and
attaches them to the echo server running in the appropriate lte service
(`lte`, `policydb`, or `subscriberdb`). We also add the proper labels and
annotations to the service registry to ensure that obsidian will reverse
proxy to the appropriate service.

Differential Revision: D22544902

fbshipit-source-id: 803599f3936ed001f7fd4ae301280d4874632e38