Attach testcontroller obsidian handlers to service itself
Summary:
This diff removes the obsidian handlers from the fbinternal
plugin and attaches them to the echo server running in the testcontroller service.
We also add the proper labels /magma/v1/ci and /magma/v1/tests to the service
registry to ensure that obsidian will reverse proxy to the testcontroller service.

Reviewed By: xjtian

Differential Revision: D22544900

fbshipit-source-id: d56875e76688fee5cca1fa34c07ba2626f98e714