Began styling global dialogs using eNodeB detail edit/add dialog (#2010)
Summary:
# Overview
This PR focuses on applying global styles for the various dialogs we use. At the moment this PR is focusing solely on eNodeB detail dialogs, but the global styles are applied throughout magma. Future PR's will align other modals to the design.

## Dialog Modal
The global styles for the dialog box itself, including `Backdrop`, `DialogTitle`, `DialogContent`, and `DialogActions` have been added to our default styles and will propagate across the platform. In addition, I've added another option for users to close the modal via a close button in the top right corner.

Input fields have had their styles updated as well to better match that of the design. I've added a new `focus` and `hover` state, as well as removed some of the curve to the corners.

#### Before
![image](https://user-images.githubusercontent.com/8878152/88107079-84a0d880-cb74-11ea-8cfa-07f0ea32f8b5.png)

#### After
![image](https://user-images.githubusercontent.com/8878152/88107027-6fc44500-cb74-11ea-80b5-30a5d45142a2.png)

Pull Request resolved: https://github.com/facebookincubator/magma/pull/2010

Reviewed By: xjtian

Differential Revision: D22669134

Pulled By: karthiksubraveti

fbshipit-source-id: fbe17c71fe1d5b7f50cabfac49971b7bc46a6450