[AGW] GTP: OVS patches updated to OVS 2.14
GTP patches are updates to OVS 2.14, This allows up to get
rid of bunch of patches that were backported from OVS master for
supporting kernel 5.6.

This is based of commit 42f667e223c005683185a97dd092545d27f29a04
from branch-2.14.

This patch also remove support for OVS 2.9 and 2.13 which is not
going to be supported.

Signed-off-by: Pravin B Shelar <pbshelar@fb.com>