[nms] remove transitive dependencies from flow (#2767)
By doing this, we reduce the number of files flow has to parse
from ~40k to ~27k.  Since they are used in package-of-packages,
and we never import those directly, we can safely ignore them

run `flow` before and after this change.  Verify no flow errors, and
that parsing has sped up

Signed-off-by: Josh Braegger <jbraeg@fb.com>