AGW: mobilityD: Do not allocate IP in case of static IP allocation error
MobilityD tries to allocate static IP, it communicates with
subscriberDB. If there is any error in getting the info from
subscriberDB, it continue with ip-pool based allocation.
This does not work when operator wants to only use static IPs
this patch changes this behaviour to failed allocation in
such cases.

Signed-off-by: Pravin B Shelar <pbshelar@fb.com>