[sctpd] Fix pointer inequality comparison with zero #clang
In effort to bring up Clang compilation (Issue #4865) - aside from a barrage of new warnings, there are some errors being found.

Clang reported here:

```shell
/home/vagrant/magma/lte/gateway/c/oai/tasks/sctp/sctpd_downlink_client.cpp:137:67: error: ordered comparison between pointer and zero ('const char *' and 'int')
    if (inet_ntop(AF_INET, &ipv4_addr, ipv4_str, INET_ADDRSTRLEN) < 0) {
        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^ ~
/home/vagrant/magma/lte/gateway/c/oai/tasks/sctp/sctpd_downlink_client.cpp:146:69: error: ordered comparison between pointer and zero ('const char *' and 'int')
    if (inet_ntop(AF_INET6, &ipv6_addr, ipv6_str, INET6_ADDRSTRLEN) < 0) {
        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^
```

It is unclear to me how this would behave in GCC.  My best guess would be that no error checking (for null pointer) is occuring.

In this PR I've moved this comparison to `!= nullptr` instead of `< 0`.

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>