Fix potential null pointer dereference in log.c
It's unclear to me whether is_async can change over the course of
execution in e.g. log_message_int_prefix_id or log_message_int.

But if this is possible (and static analysis wouldn't know if it
wasn't), then we can pass through the top check of g_oai_log.is_async
and assume it is false -> not initialize async behavior. Then hit an
error in e.g. the bvcformata call and goto error_event -> where it is
conceivable that is_async has toggled to true -> and we now de-reference
it's NULL-initialized value.

Here we add a guard against this sequence of events, however unlikely
given the use pattern of logging (I would note that the g_oai_log
structure is statically initialized, I'm not certain there are
guarantees on the ordering of its initialization and first function
calls?).

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>