Free allocated cid after use
clang-tidy currently shows a warning in the file
mme_app_edns_emulation.c due to an allocated cid
value that is never freed. Free that value prior
to returning and right after its last use. This
approach frees the CID value both in cases when
the hashtable insertion works as expected and when
it fails.

Signed-off-by: Anusha Datar <anushadatar@gmail.com>