[CI] Add CodeCov.io flags to disambiguate coverage reports (#5788)
We are observing erratic coverage reporting as CodeCov.io wants every git hash being reported to have complete coverage statistics - which we are not doing for latency / CPU / cost reasons (e.g. if a PR does not touch c/c++ targets we are omtting the hour+ of testing and upload of coverage).  This PR attempts to use per-coverage-domain tags to alleviate the issue.

The strategy is to apply a codecov.io project flag per CircleCI pipeline stage that pushes a coverage report - to ensure each can be run in isolation of any others.

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>

Co-authored-by: Scott Harrison Moeller <smoeller@fb.com>