[DevExp] Add dockerfile health tester (#6011)
## Summary

Once this GH Action has been in place for awhile, we can turn up `Required` status for the following and prevent Dockerfile breakages such as those I committed in #5939.

- build_ubuntu20_04_dockerfile
- build_devcontainer_dockerfile

## Take Note of Caching Change

I am attempting to change the caching behavior for dockerx builds in GH.  Please take a look. The intent here is to re-use the buildx cache so long as the dockerfiles repo-wide are unchanged.  We could additionally insert date strings here to force rollover on some weekly cadence...

```yaml
        with:
          path: /tmp/.buildx-cache
          key: ${{ runner.os }}-buildx-${{ hashFiles( '**/Dockerfile*' ) }}
```

## Test

Cherry-picked my master-breaking PR #5939 and observed failure to build one of our two docker containers tested in this PR.

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>

Co-authored-by: Scott Harrison Moeller <smoeller@fb.com>