Rebuild common services using a super build
Today we implicitly build the common libraries vs importing them as an
external project from the core and session manager projects.
Create a super build makefile that builds each of the libraries as an
exported package that can be imported by other binaries that depend on it.

This change doesn't change the scope of the compilation. It is likely that
we should have fewer .a's but this change doesn't change the scope of
compilation.

This change also renames the CONFIG module to MAGMA_CONFIG to allow
find_package to find MAGMA_CONFIG vs libconfig

Signed-off-by: Amar Padmanabhan <amarpadmanabhan@fb.com>