Fix SessionManager ninja cyclic dependency
Ninja builds generate a dependency graph and when the build is rerun
as part of make coverage it fails due to a cyclic dependency on common.pb

lib datastore->depends->common.pb due to redis_client
session_manager->depends->lib-datastore and thus common.pb
session_manager->depends on common.pb due to service303 and friends.

Fix this by removing the protobuf dependency at session_manger level
and instead relying on datastore to pull in the dependency.
Same logic applies for policydb as well.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>