Currently service303 reaches across service boundaries to read (#7710)
mme_app state. Fix this by pushing the stats over an ITTI message
Also remove the check for MME service as the timer is not started
in the SPGW context (as it is only run in the MME_APP)

Also may close #7421 where the nas datastructure is null as there is probably
races with clearing out the state_p.

Also, addresses race in initialization where the timer was started outside
mme_app_thread thus the zmq context might not be initialized

This reverts commit ab6d9cbe3dcd403d569df6b9142fe5c181e240ca.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>
Signed-off-by: Ramon Melero <ramonmelero@fb.com>