Move stats collection of eNB stats into s1ap thread
The current access pattern violates the ITTI framework so
1 - Create a stats timer for S1ap to send stats to service303
2 - Remove fields from mme_app structure relating to eNB (s1).
3 - Move the increment of enb stats to post s1setup request vs when the
association gets enabled. This makes more sense and maps to where the
increment used to happen for the cross thread stats
Also was missing a break for handling APPLICATION_MME_APP_STATS_MSG which
triggered the premature termination of the service303 thread, regression
introduced in by me in 118e113b35. Fixed that as well.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>