feat(agw): Pass mconfig and service config to post_processing
The PostProcessor class is the class where model specific customization
happens. The current interface to this class is too simple and assumes the
post processing is always the same. Extend it to include configs passed into
the eNB. This allows other eNB models (like Sercomm/FreedomFiOne) to
customize the eNB settings based on the config.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>