docs(or8r): Add Life of a Sync RPC initial documentation
In working to address an ongoing SEV related to Sync RPC, documentation is broadly lacking so this PR is an attempt to improve that situation.

This documentation will be reconstructed from bits of knowledge on the team, and through digging in code (as nobody still here at Magma knows this subsystem in total detail).

Signed-off-by: Scott Harrison Moeller <smoeller@fb.com>