fix(agw): Add support for ParameterKey
ParameterKey is a field that can be used to communicate the generation of
config that is being set on the enodeb. This is useful for debugging and
some models of eNB (e.g. Sercomm) require this to be set.
Also fix the type of ParameterKey to be correct in the model definition
(per standard) and updates tr-69 tests to validate Key and NS.
Since there is no way to validate that ParamKey is supported by other
eNBs add the field through an opt in flag.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>