fix(agw) Update WaitEmptyMessageState logic
Add a get_message implementation to WaitEmptyMessageState that returns
and empty Dummy message. This allows for the ACS state to transition to
it even in response to any initial message like bootstrap.

Also fix the if logic in the read_message to return unknow_param_transition
only if the method is set.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>