feat(agw): Cherry pick FreedomFiOne enodebd changes (#9056)
* fix: Remove as_dict override for spyne (#8718)

The underlying memory leak has been fixed in spyne so remove
override

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Handle empty ctx descriptor for soap messages (#8739)

In certain cases like empty HTTP ok messages the ctxt descriptor itself
might be None. Add a check for this case before dereferencing out_message


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Rename Reboot state to be generic (#8740)

The reboot command is the same across eNBs so rename it to reflect
the same

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): Pass mconfig and service config to post_processing (#8772)

The PostProcessor class is the class where model specific customization
happens. The current interface to this class is too simple and assumes the
post processing is always the same. Extend it to include configs passed into
the eNB. This allows other eNB models (like Sercomm/FreedomFiOne) to
customize the eNB settings based on the config.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Add support for ParameterKey (#8774)

ParameterKey is a field that can be used to communicate the generation of
config that is being set on the enodeb. This is useful for debugging and
some models of eNB (e.g. Sercomm) require this to be set.
Also fix the type of ParameterKey to be correct in the model definition
(per standard) and updates tr-69 tests to validate Key and NS.
Since there is no way to validate that ParamKey is supported by other
eNBs add the field through an opt in flag.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Handle WaitSetParameterValue correctly (#8850)

* fix(agw): Handle WaitSetParameterValue correctly

ACS clients can return non zero values for set calls on success (per std).
Make an optional flag to allow for SetParams to return non zero values.
Validated against FreedomFiOne and Baicells which return 1, 0 for
SetParameterValueResponse message.Status respectively.
Make change backwards compatible by using optional flag.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw) Add a unit tests for WaitSetParameterValuesState

Add a unit test
Also creates a couple of dummy classes in case we want to extend the
test coverage to other state classes as well

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Update WaitEmptyMessageState logic (#8857)

* fix(agw) Update WaitEmptyMessageState logic

Add a get_message implementation to WaitEmptyMessageState that returns
and empty Dummy message. This allows for the ACS state to transition to
it even in response to any initial message like bootstrap.

Also fix the if logic in the read_message to return unknow_param_transition
only if the method is set.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Add unit test for WaitEmptyMessage

Add test to catch missing_param transition issue and also generic coverage

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Fix formatting suggestions

Fix some of the formatting suggestions to make code cleaner

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Misc minor eNB fixes (#8979)

Some minor eNBd fixes
1 - Fix log file location
2 - Only type convert GPS Status if it is not already bool
3 - Handle PUTs and POST, sercomm eNB reports stats as PUTS and Baicells
as POST.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): Add support for new managed eNB: FreedomFiOne (#9003)

* Add CBRS fields to enodebd.yml to allow for field trials

Read CBRS SAS parameters from yml file while domain
proxy is under development.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): patch(0/4) Add support for FreedomFiOne eNBs

Expose use_param_key through BasicEnodebAcsStateMachine

This allows derived classes to customize behavior

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): patch(1/4) Add support for FreedomFiOne eNBs

The current configuration_init is pretty Baicells specific.
Only set common parameters unconditionally in configuration_init

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): patch(2/4) Add support for FreedomFiOne eNBs

Allow for stateful datamodel objects that don't exist in the eNB datamodel

Introduce a new TRparam path for such nodes.
Example usage would be PTP_STATUS that is True IFF GPS is in sync and
SyncStatus is True

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): patch(3/4) Add support for FreedomFiOne eNBs

Introduce a new FreedomFi One eNB device model

Currently SAS fields are read from enodebd.yml. This will be updated
to a periodic poll loop once DP is fully implemented.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): patch(4/4) Add support for FreedomFiOne eNBs

Add unit tests to cover provisioning, rebooting and enodebd.yml parsing

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* feat(agw): Adress review comments

Update node name from web_ui_enable -> web_ui_enable_list

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Update some formatting glitches

Ran linter

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

* fix(agw): Fix freedomfiOne unit tests (#9063)

The order in which the desired configuration is set is based on a
dictionary. This results in the call order not being stable.
Sort the list of calls before comparing with existing results.
Since the call to the eNB itself is a single call the order in which
each param is set is not relevant, so it doesn't make sense to use
the more expensive OrderedDict datastructure.

Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>

Co-authored-by: Amar Padmanabhan <amar@freedomfi.com>