ci: adds scheduled GH Action docker builds for dev / CI (#9182)
Working toward #9178.

The Action builds the vscode devcontainer Dockerfile and pushes it to the GH Container Registry. This allows GH Code Spaces to rapidly spin up by local registry access of the devcontainer.  This PR may be extended in follow-up to include other CI related docker images (e.g. the MME build container).

The GitHub Action is triggered on `Push` that modifies the associated Dockerfile, and also executes on a fixed schedule four times a day.

Pushed to a novel branch of `magma/magma` and confirmed build and push to Github Container Registry, then used the container in GitHub Code Spaces with #9177. Validated successful codespace loading / operation.

Signed-off-by: Scott Moeller <electronjoe@gmail.com>
Signed-off-by: salman-momin07 <salman.momin@wavelabs.ai>