ci: accelerates the bazel workflows by pulling from GHCR (#9229)
## Summary

### Impact

Bazel build times

- 27 minutes baseline
- 24 minutes With GHCR Pull
- 3 minutes with GHCR Pull + Bazel Cache (includes ~1 minute of dockerfile fetch)

Bazel build + test

<img width="650" alt="Screen Shot 2021-09-22 at 10 33 09 AM" src="https://user-images.githubusercontent.com/3752618/134365478-1e114c8c-c8d0-4238-b65c-63cbd8570cdc.png">

### Approach

PR overlays two GH Caches (actions/cache@v2) onto the docker run (addnab/docker-run-action@v2) - one for the Bazel Repo cache (storing external deps), one for the Bazel cache (storing build artifacts).

I found I had to provide a unique cache key (so it will be missing) - I don't know why and maybe this needs to be re-verified. Advice for compensation on this is to define `restore-key` options so that they wildcard-fallback to the prior cache that exact matches the `restore-key` or is a longest match on `restore-key`. Here we rely on longest match.

GH Cache is limited to 5 GB per repo - and is evicted LRU manner. I think we should drop our use of this cache for docker buildx behavior as they are likely consuming all of this cache - and instead use it for high gain / smaller storage e.g. Bazel and others.

The present PR caches ~500 MB of external repo contents in the bazel-cache-repo and ~210 MB of build artifacts (this will grow with time - as Bazel attempts to cache all of history).

### Open Questions

1. Do we want to limit cache size? Can that be done with e.g. `bazelrc` config?
2. Is the key behavior used in this PR going to be effective for maximally effective caching? Do we need a differing format to give good behavior on restore-key fallback?
3. Is the key behavior used in this PR going to thrash our cache 5GB allocation? If indeed I need to use a unique key then I'm in a pickle?

## Test Plan

Measured build times with and without cache - observed above speedups.

Signed-off-by: Scott Moeller <electronjoe@gmail.com>