fix(agw): Compensate for eNB firmware param inversion
A regression was introduced into the firmware build for Sercomm where
the flag for cell reserved was inverted. Compensate for that by
setting the cell reserved bit to be True on initial connect.

At a later point when we have a software version where the issue is fixed
we can add this as a conditional check.


Signed-off-by: Amar Padmanabhan <amar@freedomfi.com>