fix(monitord): Register operational states callback correctly
Summary

Pass function call instead of function return value as callback.

Context

Instead of registering a function as callback, the code passed
the return value of the called function to the callback. This
led to an error whenever GetOperationalStates was called for
monitord. It is one of the most frequent issues in Sentry when
it is enabled for Python services.

sentry.io/organizations/lf-9c/issues/2770941608/?project=5759591

Test Plan

Executed GetOperationalStates locally.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>
Signed-off-by: Marie Bremner <marwhal@fb.com>
Signed-off-by: Jan Heidbrink <jan.heidbrink@tngtech.com>