chore(agw): Exclude message patterns in Sentry client
Summary

Currently, a lot of errors are excluded from monitoring in the
Sentry server, leading to 100,000s of filtered messages per day.
Instead, we add the possibility to configure such a filter in
the client. For a start, the client will use the same filters as
the server.

Test Plan

The filter functionality was tested with a Sentry test account
using error logs, logged exceptions and explicit captured
messages.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>
Signed-off-by: Alexander zur Bonsen <alexander.zur.bonsen@tngtech.com>