fix(agw): migrate orc8r/gateway/c to clang-11 google style (#10695)
Work to accomplish #6187.  This will be applied on a per-directory basis, with the `oai/mme` sub-directory reserved for reformatting until last (after community notification in ~early January).  This PR atomically does the following:

- Re-formats all of `/orc8r/gateway/` with `clang-format-11` and `Google` style
- Updates GH Workflow for CI to match this migration
- Updates the `lte/gateway/Makefile` target for `format_all` to support this migration

Should outstanding PRs have issues with the new CI phase / rebasing on this PR, it is advised they re-format their modified files (from within either the Magma Developer VM or the Devcontainer) with:

```bash
/usr/bin/clang-format-11 --style='{BasedOnStyle: Google, IncludeBlocks: Preserve, SortIncludes: false}' -i path/to/file1.cpp path/to/file2.h ...
```

## Test Plan

To confirm that the CI phases are working, this PR was test-modified with formatting non-compliant changes to c++ files in either `orc8r/gateway` and `lte/gateway` to confirm clang-format differences are caught per sub-directory (see #10712).

Signed-off-by: Scott Moeller <electronjoe@gmail.com>