feat(sctpd): Enable mconfig and Sentry monitoring
Summary

Introduce mconfig for SCTPd, and use it to set the log level and
Sentry configuration. Enable Sentry if a URL is given.

This uses the default log level INFO, which is also used for the
other services, instead of DEBUG as hardcoded before.

Test Plan

Sent a message to Sentry test account via local SCTPd after
configuring the Sentry URL in Swagger.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>