fix(ci): Check out full history in build for Sentry release tag
Summary

Changes the GHA checkout of the repository to fetch all commits.

The Magma version in the new Sentry release tag needs to
calculate the number of commits for the current release. To do
this it has to check the history and cannot work with a shallow
clone of the repository.

Test Plan

I did a shallow clone of the repository locally via:

git clone --depth 1 git@github.com:magma/magma.git

This reproduces the error we see in Sentry where the commit hash
is correct but the commit count is 1.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>