feat(agw): Implement opt-out for Sentry, remove other options
Summary

As an intermediate solution with maximum configurability, Sentry
could be disabled or enabled per Python service and run in a
special opt-in mode that only tracked a few error log messages.
After enabling Sentry in the HIL, we got a better estimation of
the logging volume and issue, and these options were only
rarely needed.

To reduce complexity, we remove the Sentry options again and
replace them by an opt-out tag that can be used to exclude error
logs from Sentry monitoring.

See the discussion in #10027 for more details.

Test Plan

Added unit tests and executed a manual end-to-end test with a
Sentry test account.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>