feat(ci): activate more diverse docker image tagging
Within #10711 the foundation was laid and an argument was made for a more diverse docker image tagging. However, it turns out that only the `latest` tag was indeed introduced.
The default values implemented in #10711 are currently overwritten by workflow specific ones. Those workflow specific overwrites are deleted in this PR to fully activate the more diverse image tagging.

Signed-off-by: Fritz Lehnert <Fritz.Lehnert@tngtech.com>