chore(mme): Commits failing Bazel core oai tests with manual tag
This PR commits the final four oai core tests as Bazel test targets, but
as each of them fail in some manner across `vanilla`, `asan` or `lsan`
configurations we use the `manual` tag so that they are not run in CI.

Manual tag use is not best practice generally, as it can lead to hidden
tests that are broken and don't get attention. But since we have a
focused team working on fixing these tests - and it seems valuable to
have the team working on master and not my branch - here we propose
committing them with a `manual` tag.

Validated that we do not break any Bazel build or test with this PR.
Validated that tests can be run manually and fail per their GH Issues.

Signed-off-by: Scott Moeller <electronjoe@gmail.com>