fix(mme): Provide cpp compatible variant of PARENT_STRUCT
Closes #11585.

The existing macro fails to compile In Bazel build effort because some paths to the containing header
are not guarded by `extern "C"`. This is a common issue with the Magma
code-base and quite challenging to fixup. It will be aided by rename of
CPP headers to .hpp (see GH issue associated with that effort).

I did attempt to fixup all the missing `extern "C"` calls, but then ran
into new errors on the macro compilation that appeared again related to
extern "C" use.

Test Plan

CI tests include `make test_oai` and `make build_oai`.

Signed-off-by: Scott Moeller <electronjoe@gmail.com>