feat(dp): Do not send amc config for cbsds with pending requests
Previously RC sent pending requests to AMC and AMC filtered pending
requests itself.

However there were a few issues with that approach:
- AMC just compared requests directly, but it no longer works,
because of introduction of randomness in channel selection,
which completely invalidates the current approach,
- for given cbsd AMC can only generate one request at a time,
and when some request was pending AMC generated the same one,
so this was just some unnecessary work,
- when AMC wanted to delete cbsd it didn't even need requests content,
it just needed information if some requests are pending.

Now AMC do even receive any data for cbsds,
and because of the fact, that AMC could only send one request at a time,
the functional behavior remains unchanged.

Signed-off-by: Kuba Marciniszyn <kuba@freedomfi.com>