chore(dp): Unify null and non null types behavior on insert
Previously null and non null types were treated differently on insert.
For example when given field didn't have value and it was required,
then default value was used instead of null.

But there is no point of having separate behavior, since:
- when field is not required then it can be null,
so it should be possible to assign null to it
- when field is required then it is better to explicitly fail,
when no value is provided, rather than using implicit default value

Explicit default values can still be used, with field masks.
In order to use them one has to just omit the given field in the field
mask.

Signed-off-by: Kuba Marciniszyn <kuba@freedomfi.com>