docs: fix "howtos/pcap" broken link in sidebar (#12521)
The `Next` button on the `howtos/inbound_roaming` page was linking to
a non-existent `howtos/pcap` page, resulting in `404` error.

The reason was the `howtos/pcap` entry in the `sidebars.json`, which was
added in commit `f23a49653abf82c1e1a2ec8b89dd5cda3f0fcd5f`. That commit
added the `resources/ref_pcap.md` file, and a proper link to it in the
`Technical Reference` section of the sidebar. However the `howtos/pcap`
link was also added, probably by mistake, as `howtos/pcap.md` doesn't
exist.

Delete the `howtos/pcap` link in the sidebar to fix the `Next` button
for the preceding `howtos/inbound_roaming` docs page.
The bug was present both in the `v1.7` and `v1.6`, so fix both.

Signed-off-by: Ivan Sergiienko <ivan@freedomfi.com>