Add reindexing of gateway_subscriber_state to subscriberdb indexer
Modifies the existing indexer servicer by adding the subscriber
storage table to it and functionality to take states of type
"gateway_subscriber_state" (sent by sessiond) and write an entry
for each subscriber into the subscriber storage.

For that we modify the existing indexImpl function to distinguish
between the two possible types ("mobilityd_ipdesc_record" and
"gateway_subscriber_state").

Tests for the new indexer functionality are added.
A custom serde type is registered for the gateway subscriber state
and implemented in storage/subscriber_state.go.

The existing interface of the subscriber storage is changed to use
this new type (GatewaySubscriberState).

Signed-off-by: Alexander zur Bonsen <alexander.zur.bonsen@tngtech.com>